move `ConfirmationModal` to `/common/modals`
