+++ clients/admin-ui/src/features/configure-consent/VendorCookieTable.tsx
@@ -25,8 +25,8 @@ import {
 } from "react-table";
 
 import { useAppSelector } from "~/app/hooks";
-import ConfirmationModal from "~/features/common/ConfirmationModal";
 import { getErrorMessage, isErrorResult } from "~/features/common/helpers";
+import ConfirmationModal from "~/features/common/modals/ConfirmationModal";
 import { PaddedCell } from "~/features/common/table";
 import GroupedTableBody from "~/features/common/table/grouped/GroupedTableBody";
 import GroupedTableHeader from "~/features/common/table/grouped/GroupedTableHeader";
