+++ clients/admin-ui/src/features/user-management/PermissionsForm.tsx
@@ -8,7 +8,6 @@ import {
   useDisclosure,
   useToast,
 } from "@fidesui/react";
-import ConfirmationModal from "common/ConfirmationModal";
 import { useHasPermission } from "common/Restrict";
 import { Form, Formik } from "formik";
 import NextLink from "next/link";
@@ -16,6 +15,7 @@ import React, { useEffect, useState } from "react";
 
 import { useAppSelector } from "~/app/hooks";
 import { getErrorMessage, isErrorResult } from "~/features/common/helpers";
+import ConfirmationModal from "~/features/common/modals/ConfirmationModal";
 import { USER_MANAGEMENT_ROUTE } from "~/features/common/nav/v2/routes";
 import QuestionTooltip from "~/features/common/QuestionTooltip";
 import { errorToastParams, successToastParams } from "~/features/common/toast";
